Add camera configuration

+++ Engine/Frames.cpp
@@ -42,7 +42,7 @@ void Frames::_Render()
 	if (Camera::GetMainCamera() != nullptr)
 	{
 		m_FpsDisplay.SetX(static_cast<int>(Camera::GetMainCamera()->GetXOffset()) + 2);
-		m_FpsDisplay.SetY(Camera::GetMainCamera()->GetYOffset() + 2);
+		m_FpsDisplay.SetY(static_cast<int>(Camera::GetMainCamera()->GetYOffset()) + 2);
 	}
 	m_FpsDisplay.Render();
 }
